requirements.txt
pymupdf
reportlab
google-generativeai
pytesseract
pdf2image
pillow
